pyproject.toml
[tool.poetry]
name = "deejeiango"
version = "0.1.0"
description = ""
authors = ["Jonggwan Park <craftguy55@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
ruff = "^0.6.1"
pyjwt = "^2.9.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
kafka-python = "^2.0.2"
celery = "^5.4.0"
requests = "^2.32.3"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
